+++ android-components/components/browser/menu/src/main/java/mozilla/components/browser/menu/item/WebExtensionBrowserMenuItem.kt
@@ -140,8 +140,7 @@ class WebExtensionBrowserMenuItem(
                 Log.Priority.ERROR,
                 "mozac-webextensions",
                 throwable,
-                "Failed to load browser action icon, falling back to default.",
-            )
+            ) { "Failed to load browser action icon, falling back to default." }
 
             getDrawable(context, iconsR.drawable.mozac_ic_web_extension_default_icon)
         }