+++ android-components/components/feature/toolbar/src/main/java/mozilla/components/feature/toolbar/WebExtensionToolbarAction.kt
@@ -85,8 +85,7 @@ open class WebExtensionToolbarAction(
                     Log.Priority.ERROR,
                     "mozac-webextensions",
                     throwable,
-                    "Failed to load browser action icon, falling back to default.",
-                )
+                ) { "Failed to load browser action icon, falling back to default." }
             }
         }