+++ android-components/samples/browser/src/gecko/java/org/mozilla/samples/browser/Components.kt
@@ -33,11 +33,19 @@ class Components(private val applicationContext: Context) : DefaultComponents(ap
         GeckoEngine(applicationContext, engineSettings, runtime).also {
             it.installBuiltInWebExtension("borderify@mozac.org", "resource://android/assets/extensions/borderify/") {
                     throwable ->
-                Log.log(Log.Priority.ERROR, "SampleBrowser", throwable, "Failed to install borderify")
+                Log.log(
+                    priority = Log.Priority.ERROR,
+                    tag = "SampleBrowser",
+                    throwable = throwable,
+                ) { "Failed to install borderify" }
             }
             it.installBuiltInWebExtension("testext@mozac.org", "resource://android/assets/extensions/test/") {
                     throwable ->
-                Log.log(Log.Priority.ERROR, "SampleBrowser", throwable, "Failed to install testext")
+                Log.log(
+                    priority = Log.Priority.ERROR,
+                    tag = "SampleBrowser",
+                    throwable = throwable,
+                ) { "Failed to install testext" }
             }
             WebCompatFeature.install(it)
             WebCompatReporterFeature.install(it)