+++ fenix/app/src/test/java/org/mozilla/fenix/FenixLogSinkTest.kt
@@ -6,6 +6,7 @@ package org.mozilla.fenix
 
 import io.mockk.spyk
 import io.mockk.verify
+import mozilla.components.support.base.log.Log
 import mozilla.components.support.base.log.sink.AndroidLogSink
 import org.junit.Before
 import org.junit.Test
@@ -21,22 +22,26 @@ class FenixLogSinkTest {
 
     @Test
     fun `GIVEN we're in a release build WHEN we log debug statements THEN logs should not be forwarded`() {
+        Log.reset()
         val logSink = FenixLogSink(false, androidLogSink)
-        logSink.log(
-            mozilla.components.support.base.log.Log.Priority.DEBUG,
-            "test",
-            message = "test",
+        Log.addSink(logSink)
+        Log.log(
+            priority = mozilla.components.support.base.log.Log.Priority.DEBUG,
+            tag = "test",
+            message = { "test" },
         )
         verify(exactly = 0) { androidLogSink.log(any(), any(), any(), any()) }
     }
 
     @Test
     fun `GIVEN we're in a release build WHEN we log error statements THEN logs should be forwarded`() {
+        Log.reset()
         val logSink = FenixLogSink(false, androidLogSink)
-        logSink.log(
+        Log.addSink(logSink)
+        Log.log(
             mozilla.components.support.base.log.Log.Priority.ERROR,
             "test",
-            message = "test",
+            message = { "test" },
         )
 
         verify(exactly = 1) {
@@ -50,11 +55,13 @@ class FenixLogSinkTest {
 
     @Test
     fun `GIVEN we're in a release build WHEN we log warn statements THEN logs should be forwarded`() {
+        Log.reset()
         val logSink = FenixLogSink(false, androidLogSink)
-        logSink.log(
+        Log.addSink(logSink)
+        Log.log(
             mozilla.components.support.base.log.Log.Priority.WARN,
             "test",
-            message = "test",
+            message = { "test" },
         )
         verify(exactly = 1) {
             androidLogSink.log(
@@ -67,11 +74,13 @@ class FenixLogSinkTest {
 
     @Test
     fun `GIVEN we're in a release build WHEN we log info statements THEN logs should be forwarded`() {
+        Log.reset()
         val logSink = FenixLogSink(false, androidLogSink)
-        logSink.log(
+        Log.addSink(logSink)
+        Log.log(
             mozilla.components.support.base.log.Log.Priority.INFO,
             "test",
-            message = "test",
+            message = { "test" },
         )
         verify(exactly = 1) {
             androidLogSink.log(
@@ -84,11 +93,13 @@ class FenixLogSinkTest {
 
     @Test
     fun `GIVEN we're in a debug build WHEN we log debug statements THEN logs should be forwarded`() {
+        Log.reset()
         val logSink = FenixLogSink(true, androidLogSink)
-        logSink.log(
+        Log.addSink(logSink)
+        Log.log(
             mozilla.components.support.base.log.Log.Priority.DEBUG,
             "test",
-            message = "test",
+            message = { "test" },
         )
 
         verify(exactly = 1) {